Small fix for integer overflow for big perft runs

+++ UCIEngine/Core/Engine/MoveGen/Verification.cs
@@ -39,7 +39,7 @@ namespace Chess
             for (int i = 1; i <= depth; i++)
             {
                 Stopwatch timer = Stopwatch.StartNew();
-                int numNodes = Perft(i);
+                long numNodes = Perft(i);
                 timer.Stop();
                 TimeSpan timespan = timer.Elapsed;
                 Console.WriteLine("Depth " + i + " ply: " + numNodes + " nodes found in " + String.Format("{0:00} minutes {1:00} seconds {2:00} milliseconds", timespan.Minutes, timespan.Seconds, timespan.Milliseconds));
@@ -51,12 +51,12 @@ namespace Chess
         /* Perft (performance test, move path enumeration). This function serves a debugging function designed to walk the move generation
         tree of legal moves to count all leaf nodes to a certain depth. These values can be compared to predetermined values in order to
         isolate issues with movegen. */
-        public static int Perft(int depth)
+        public static long Perft(int depth)
         {
             if (depth == 0) return 1;
 
             Move[] moves = GenerateMoves();
-            int numPositions = 0;
+            long numPositions = 0;
             for (int i = 0; i < moves.Length; i ++)
             {
                 ExecuteMove(ref moves[i]);
@@ -70,12 +70,12 @@ namespace Chess
         public static void DebugPerft(int depth)
         {
             Move[] moves = GenerateMoves();
-            int total = 0;
+            long total = 0;
 
             for (int i = 0; i < moves.Length; i++)
             {
                 ExecuteMove(ref moves[i]);
-                int nodes = Perft(depth - 1);
+                long nodes = Perft(depth - 1);
                 total += nodes;
                 Console.WriteLine($"{ToUCI(ref moves[i])}: {nodes}");
                 UndoMove(ref moves[i]);
